# Conflicts:
#	services/cms/src/blocks/ExerciseTask/ExerciseTaskEditor.tsx
#	services/cms/src/blocks/ExerciseTask/IFrameEditor.tsx
#	services/cms/src/components/editors/EmailEditor.tsx
#	services/cms/src/components/editors/ExamsInstructionsEditor.tsx
#	services/cms/src/pages/_app.tsx
#	services/cms/src/services/backend/course-instances.ts
#	services/course-material/src/components/ContentRenderer/moocfi/CongratulationsBlock/index.tsx
#	services/course-material/src/components/ContentRenderer/moocfi/ExerciseBlock/ExerciseTask.tsx
#	services/course-material/src/components/ContentRenderer/moocfi/ExerciseBlock/ExerciseTaskIframe.tsx
#	services/course-material/src/components/ContentRenderer/moocfi/ExerciseBlock/GradingState.tsx
#	services/course-material/src/components/ContentRenderer/moocfi/ExerciseBlock/PeerReviewView/PeerReviewViewImpl.tsx
#	services/course-material/src/components/ContentRenderer/moocfi/ExerciseBlock/PeerReviewView/WaitingForPeerReviews.tsx
#	services/course-material/src/components/ContentRenderer/moocfi/ExerciseBlock/index.tsx
#	services/course-material/src/components/forms/SelectCourseInstanceForm.tsx
#	services/course-material/src/pages/_app.tsx
#	services/main-frontend/src/components/page-specific/global-stats/GlobalStatTable.tsx
#	services/main-frontend/src/components/page-specific/manage/courses/id/feedback/FeedbackPage.tsx
#	services/main-frontend/src/components/page-specific/submissions/id/SubmissionIFrame.tsx
#	services/main-frontend/src/pages/_app.tsx
#	services/main-frontend/src/pages/completion-registration/[courseModuleId]/redirect.tsx
#	services/main-frontend/src/pages/playground-views.tsx
#	services/main-frontend/src/services/backend/global-stats.ts
#	services/quizzes/src/components/exercise-service-views/AnswerExercise/impl-by-quiz-item-type/ChooseN.tsx
#	services/quizzes/src/components/exercise-service-views/AnswerExercise/impl-by-quiz-item-type/Essay.tsx
#	services/quizzes/src/components/exercise-service-views/AnswerExercise/impl-by-quiz-item-type/MultipleChoice.tsx
#	services/quizzes/src/components/exercise-service-views/AnswerExercise/impl-by-quiz-item-type/Scale.tsx
#	services/quizzes/src/components/exercise-service-views/AnswerExercise/impl-by-quiz-item-type/Timeline.tsx
#	services/quizzes/src/components/exercise-service-views/ViewSubmission/impl-by-quiz-item-type/Closed-ended-question.tsx
#	services/quizzes/src/components/exercise-service-views/ViewSubmission/impl-by-quiz-item-type/Essay.tsx
#	services/quizzes/src/components/exercise-service-views/ViewSubmission/impl-by-quiz-item-type/Matrix.tsx
#	services/quizzes/src/components/exercise-service-views/ViewSubmission/impl-by-quiz-item-type/Scale.tsx
#	shared-module/jest.config.js
#	shared-module/package-lock.json
#	shared-module/package.json
#	shared-module/packages/common/src/hooks/useClickOutside.tsx
#	shared-module/packages/common/src/img/moocfiLogoNoText.svg
#	shared-module/packages/common/src/styles/muiTheme.ts
#	shared-module/stories/DropdownMenu.stories.tsx
#	shared-module/tests/components/MessageChannelIFrame.test.tsx
#	system-tests/src/tests/peer-reviews/peer_review_utils.ts

+++ services/headless-lms/models/src/library/progressing.rs
@@ -358,14 +358,14 @@ pub async fn process_all_course_instance_completions(
     Ok(())
 }
 
-#[derive(Debug, Serialize, Deserialize, PartialEq, Eq, Clone)]
+#[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
 #[cfg_attr(feature = "ts_rs", derive(TS))]
 pub struct CourseInstanceCompletionSummary {
     pub course_modules: Vec<CourseModule>,
     pub users_with_course_module_completions: Vec<UserWithModuleCompletions>,
 }
 
-#[derive(Debug, Serialize, Deserialize, PartialEq, Eq, Clone)]
+#[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
 #[cfg_attr(feature = "ts_rs", derive(TS))]
 pub struct UserWithModuleCompletions {
     pub completed_modules: Vec<CourseModuleCompletionWithRegistrationInfo>,
@@ -375,7 +375,7 @@ pub struct UserWithModuleCompletions {
     pub user_id: Uuid,
 }
 
-#[derive(Debug, Serialize, Deserialize, PartialEq, Eq, Clone)]
+#[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
 #[cfg_attr(feature = "ts_rs", derive(TS))]
 pub struct UserCourseModuleCompletion {
     pub course_module_id: Uuid,
@@ -451,7 +451,7 @@ pub async fn get_course_instance_completion_summary(
     })
 }
 
-#[derive(Debug, Serialize, Deserialize, PartialEq, Eq, Clone)]
+#[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
 #[cfg_attr(feature = "ts_rs", derive(TS))]
 pub struct TeacherManualCompletionRequest {
     pub course_module_id: Uuid,
@@ -459,7 +459,7 @@ pub struct TeacherManualCompletionRequest {
     pub skip_duplicate_completions: bool,
 }
 
-#[derive(Debug, Serialize, Deserialize, PartialEq, Eq, Clone)]
+#[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
 #[cfg_attr(feature = "ts_rs", derive(TS))]
 pub struct TeacherManualCompletion {
     pub user_id: Uuid,
@@ -542,7 +542,7 @@ pub async fn add_manual_completions(
     Ok(())
 }
 
-#[derive(Debug, Serialize, Deserialize, PartialEq, Eq, Clone)]
+#[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
 #[cfg_attr(feature = "ts_rs", derive(TS))]
 pub struct ManualCompletionPreview {
     pub already_completed_users: Vec<ManualCompletionPreviewUser>,
@@ -550,7 +550,7 @@ pub struct ManualCompletionPreview {
     pub non_enrolled_users: Vec<ManualCompletionPreviewUser>,
 }
 
-#[derive(Debug, Serialize, Deserialize, PartialEq, Eq, Clone)]
+#[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
 #[cfg_attr(feature = "ts_rs", derive(TS))]
 pub struct ManualCompletionPreviewUser {
     pub user_id: Uuid,
@@ -619,14 +619,14 @@ pub async fn get_manual_completion_result_preview(
     })
 }
 
-#[derive(Clone, PartialEq, Eq, Deserialize, Serialize)]
+#[derive(Clone, PartialEq, Deserialize, Serialize)]
 #[cfg_attr(feature = "ts_rs", derive(TS))]
 pub struct UserCompletionInformation {
     pub course_module_completion_id: Uuid,
     pub course_name: String,
     pub uh_course_code: String,
     pub email: String,
-    pub ects_credits: Option<i32>,
+    pub ects_credits: Option<f32>,
     pub enable_registering_completion_to_uh_open_university: bool,
 }
 
@@ -677,7 +677,7 @@ pub async fn get_user_completion_information(
     })
 }
 
-#[derive(Clone, PartialEq, Eq, Deserialize, Serialize)]
+#[derive(Clone, PartialEq, Deserialize, Serialize)]
 #[cfg_attr(feature = "ts_rs", derive(TS))]
 pub struct UserModuleCompletionStatus {
     pub completed: bool,
@@ -755,7 +755,7 @@ pub async fn get_user_module_completion_statuses_for_course_instance(
     Ok(course_module_completion_statuses)
 }
 
-#[derive(Clone, PartialEq, Eq, Deserialize, Serialize)]
+#[derive(Clone, PartialEq, Deserialize, Serialize)]
 #[cfg_attr(feature = "ts_rs", derive(TS))]
 pub struct CompletionRegistrationLink {
     pub url: String,
